+++ whatsnew.md
@@ -1,4 +1,4 @@
 # What's new
 
 - Some app icons in a app pair shortcut may have appeared too big. This has been fixed.
-- Improved layout of the small app icons in the app pair shortcut: the icon on the right hand side no longer collides with the small Be nice icon
+- You can now specify how the small app icons in the app pair shortcut will be arranged: horizontal and diagonal